Controllers!
Controllers are things that take some files, perform some logic on them,
and then make some files in the build_dir. For example, I'm gonna write
one realy quick called "blog", that will take the files that are stored
in _posts and, with the help of some mako templates, make a blog.

To install a controller: put it in your _controllers directory (or
anywhere really), import it in your _config.py, and add an instance of
the controller object to your `controllers` list.