Completely changed how controllers work, so they actually work now.
The user has a directory called _controllers; in it are a bunch of
cytoplasm controllers (which are just disguised python modules). To make
cytoplasm use a certain controller, the user adds a tuple to the list
"controllers" in their _config.py.

This list is evauluated by the method "build" in build.py. For each line
of the list, "build" calls the method "controllerclass" in
controllers.py.

This method imports the module of the controller, and then looks in a
special dictionary called "info" that (for now at least) all controllers
must have. Specifically, it looks up 'info["class"]', which is the class
of that type of controller. This class should be a subclass of the class
"Controllers" in controllers.py. Finally, "controllerclass" returns that
class back to "build".

"build" then applies the list of arguments that came with the controller
name in "controllers" to that class, getting a Controller object. These
arguments are probably at least the source directory the Controller will
read and the build directory it should put things in.

Finally, "build" calls this Controller object, which should make it do
whatever it wants to do.