Add .encode('utf-8', 'replace') if it's a python 2.x when rendering mako
This is such an ugly hack! But so was ASCII. Complain upstream...

Since this would break things for python 3.x, have a little if loop in
there to check the version. Yeah. This sucks. Everyone should use Python
3.

This is half of the battle; the other half has to do with the blgo
controller. Look there.