Use a multiprocessing event to coordinate the server test and the server.
Previously, there was a race condition here because the test would
occasionally proceed to urllib requests before the server was done
initializing. Then the test fails.

Instead, now cytoplasm.server.serve can optionally take as an argument
an event that it will set when it's done initializing. The server test
now waits until the event is set before proceeding to urllib requests.