Made a module as a container for temporary global variables.
This makes it possible to import the current site into the configuration
and default configuration files; in fact, it's in the default
configuration now.

Basically it allows communication between modules easily. I can't decide
if this is clever or horrible. ANyway, it's kind of neat and the thing
in cytoplasm.site reverts the global back to None afterwards anyway.

This fixes/works around issue #12.